Got it working in a case that was returning null
Added one more location to check for the IP address: in the remoteAddress field of the req.info object. This succeeded in certain cases the module was previously returning null.

Careful null checks were done on req.info before attempting to use it, to prevent an error from being thrown - the info object appears to not always be present in the req object.

For some reason checking in the try-catch block failed to provide the IP. It only worked when done separately, just prior to the return. This suggests there may be a problem with that block in certain scenarios (it might be worth testing).