Return once we find something
This address https://github.com/pbojinov/request-ip/issues/15 in that we return
once we find an address in the prescribed order. It only looks at as many elements as necessary.
This is definitely a micro-optimization but I think it helps clarity as well. This still retains the
comments as to where these headers come from.

This also address the possiblity of req.headers not existing. This might be an extreme edgecase but
it is one I have run into in my usage and having the fallback to not use the headers if they do not
exist within the library would help me.